Create MazeGameDij&A*.py
I changed the design of the program using objected-oriented approach. I implemented A* algorithm and showed how faster it is than Dijstra's algorithm by coloring visited and inspected nodes, comparing the time elapsed and number of iterations. My next step is to make different threads for GUI and algorithms and show how the best solution is found step-by-step.